requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
google-generativeai==0.3.2
faiss-cpu==1.7.4
numpy==1.24.3
PyPDF2==3.0.1